+++ packages/events/src/service/indexing/query.ts
@@ -360,6 +360,11 @@ export function withJurisdictionFilters({
               term: { placeOfEvent: value }
             })
             break
+          case 'createdBy':
+            must.push({
+              term: { createdBy: value }
+            })
+            break
           case 'declaredIn':
             must.push({
               term: {