+++ packages/events/src/service/indexing/utils.ts
@@ -474,6 +474,8 @@ export function resolveRecordActionScopeToIds(
     options: {
       event: options?.event,
       placeOfEvent: getLocationIdsFromScopeOptions(options?.placeOfEvent, user),
+      createdBy:
+        options?.createdBy === UserFilter.enum.user ? user.id : undefined,
       declaredIn: getLocationIdsFromScopeOptions(options?.declaredIn, user),
       declaredBy:
         options?.declaredBy === UserFilter.enum.user ? user.id : undefined,